requirements.txt
pandas==2.1.4
matplotlib==3.8.2
tabulate==0.9.0
psutil==5.9.7
pyyaml==6.0.1